[SaturnClient] Fixed main search bar performance regression.
I'm not entirely sure where the performance regression was introduced as a lot as been changed in SATurn since this was noticed.  Users haven't been affected as the code with the regression in hasn't been deployed to them.  Rather than sending a search request and then performing the next when the first returns we now send them in quick sucession and then wait for all to return their results.