[ChromoHubViewer] Removed JavaScript polling from the two functions which process annotations after they have been fetched.
Sefa was using a polling mechanism to check to see if all async functions had either returned an error or returned results.  This resulted in the UI lagging at times and also slows down the display of annotations.  The new implementation doesn't require a polling mechanism to detect when all async functions have completed.